Update a unit test to work with latest r-devel
The function class(), when applied to matrices, is being changed to return
a list of two items ("matrix" and "array") instead of just "matrix". One of my
unit tests used an expect_equal() to "matrix", which would now fail. I just
removed that unit test, as the two that follow should be sufficient.